% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{example}
\format{
An object of class \code{data.frame} with 3093 rows and 285 columns.
}
\usage{
example
}
\description{
Just test number(Description)
}
\examples{
  head(example)
}
\keyword{datasets}
