% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGS.R
\name{genotype_analysis}
\alias{genotype_analysis}
\title{Systematic GWAS prior information GS strategy (SGGS)}
\usage{
genotype_analysis(input_file)
}
\arguments{
\item{input_file}{Path to the input file.}
}
\value{
Returns the result of SGGS geno analysis.
}
\description{
Systematic GWAS prior information GS strategy (SGGS)
}
